plugins/joycaption_reverse_plugin/requirements.txt
torch>=2.0.0
transformers>=4.30.0
huggingface_hub>=0.16.0
Pillow>=9.0.0
requests>=2.28.0
bitsandbytes>=0.41.0
accelerate>=0.20.0
safetensors>=0.3.0 